refactor: AdminMenuCardを独立コンポーネント化
